requirements.txt
requests>=2.28.0
PyYAML>=6.0
pytest>=7.0
pytest-html>=3.2.0
pytest-cov>=4.0.0